refactor(game+api): break up events.rs, tributes/mod.rs, brains/mod.rs, timeline.rs
Extract api/src/templates/timeline.rs → timeline/mod.rs + cards.rs

Extract game/src/events.rs → events/types.rs + events/mod.rs
Extract game/src/tributes/mod.rs → helpers.rs + tests.rs (trimmed mod.rs)
Extract game/src/tributes/brains/mod.rs → scoring.rs + tests.rs (trimmed mod.rs)

No logic changes. 956 game tests pass.
