+++ game/src/tributes/helpers.rs
@@ -1,0 +1,148 @@
+//! Helper functions and types for the tributes module.
+//!
+//! Extracted from the monolithic `mod.rs` to keep the main module focused
+//! on the `Tribute` struct and its core impl.
+
+use serde::{Deserialize, Deserializer, Serializer, ser::SerializeSeq};
+use uuid::Uuid;
+
+use crate::tributes::Tribute;
+use crate::tributes::actions::Action;
+use shared::afflictions::{AfflictionKind, AfflictionSource, BodyPart, Severity};
+
+/// Serialize `Vec<Uuid>` as `Vec<String>` for SurrealDB compatibility.
+/// The Surreal Rust SDK's bespoke serializer wires `uuid::Uuid` as raw bytes,
+/// which Surreal then renders as base64 and rejects against `array<uuid>`
+/// constraints. Storing as strings on the wire (and as `array<string>` in
+/// the schema) follows the same convention as `message.event_id`.
+pub fn serialize_uuids_as_strings<S>(uuids: &[Uuid], serializer: S) -> Result<S::Ok, S::Error>
+where
+    S: Serializer,
+{
+    let mut seq = serializer.serialize_seq(Some(uuids.len()))?;
+    for u in uuids {
+        seq.serialize_element(&u.to_string())?;
+    }
+    seq.end()
+}
+
+/// Deserialize `Vec<Uuid>` from either a sequence of strings (the wire format
+/// we write) or a sequence of native uuid values (test fixtures, JSON read
+/// back through serde's standard Uuid impl).
+pub fn deserialize_uuids_lenient<'de, D>(deserializer: D) -> Result<Vec<Uuid>, D::Error>
+where
+    D: Deserializer<'de>,
+{
+    #[derive(Deserialize)]
+    #[serde(untagged)]
+    enum StringOrUuid {
+        S(String),
+        U(Uuid),
+    }
+
+    let raw: Vec<StringOrUuid> = Vec::deserialize(deserializer)?;
+    raw.into_iter()
+        .map(|item| match item {
+            StringOrUuid::S(s) => Uuid::parse_str(&s).map_err(serde::de::Error::custom),
+            StringOrUuid::U(u) => Ok(u),
+        })
+        .collect()
+}
+
+/// Serialize a single `Uuid` as a string for the same reasons as
+/// `serialize_uuids_as_strings`.
+pub fn serialize_uuid_as_string<S>(uuid: &Uuid, serializer: S) -> Result<S::Ok, S::Error>
+where
+    S: Serializer,
+{
+    serializer.serialize_str(&uuid.to_string())
+}
+
+/// Deserialize a single `Uuid` from either a string (our wire format) or the
+/// SDK's native uuid bytes representation.
+pub fn deserialize_uuid_lenient<'de, D>(deserializer: D) -> Result<Uuid, D::Error>
+where
+    D: Deserializer<'de>,
+{
+    #[derive(Deserialize)]
+    #[serde(untagged)]
+    enum StringOrUuid {
+        S(String),
+        U(Uuid),
+    }
+
+    match StringOrUuid::deserialize(deserializer)? {
+        StringOrUuid::S(s) => Uuid::parse_str(&s).map_err(serde::de::Error::custom),
+        StringOrUuid::U(u) => Ok(u),
+    }
+}
+
+/// Sanity level at which a tribute may attempt suicide.
+pub(crate) const SANITY_BREAK_LEVEL: u32 = 9;
+
+/// A draft affliction ready for acquisition resolution.
+#[derive(Clone)]
+pub struct AfflictionDraft {
+    pub kind: AfflictionKind,
+    pub body_part: Option<BodyPart>,
+    pub severity: Severity,
+    pub source: AfflictionSource,
+}
+
+/// Calculates the stamina cost for a tribute action based on:
+/// - Base action cost
+/// - Terrain movement multiplier
+/// - Terrain affinity modifier (0.8 if tribute has affinity, 1.0 otherwise)
+/// - Desperation multiplier based on health (1.0 + 0.5 * (1.0 - health%))
+pub fn calculate_stamina_cost(
+    action: &Action,
+    terrain: &crate::terrain::TerrainType,
+    tribute: &Tribute,
+) -> u32 {
+    // Base costs for each action type
+    let base_cost: f32 = match action {
+        Action::Move(_) => 20.0,
+        Action::Hide => 15.0,
+        Action::TakeItem => 10.0,
+        Action::Attack => 25.0,
+        Action::Rest | Action::None => 0.0,
+        Action::UseItem(_) => 10.0,
+        // Proposing an alliance is a low-cost social action.
+        Action::ProposeAlliance => 5.0,
+        // Survival actions: foraging/seeking shelter cost some stamina;
+        // eating and drinking are essentially free overhead.
+        Action::SeekShelter => 10.0,
+        Action::Forage => 15.0,
+        Action::DrinkFromTerrain => 5.0,
+        Action::Eat(_) | Action::DrinkItem(_) => 0.0,
+        // Sleep is free at the action layer; phase scheduler handles it.
+        Action::Sleep { .. } => 0.0,
+        // Frozen tribute takes no action — no stamina cost.
+        Action::Frozen | Action::Flashback { .. } | Action::Avoidance => 0.0,
+    };
+
+    // If base cost is 0, no need to calculate multipliers
+    if base_cost == 0.0 {
+        return 0;
+    }
+
+    // Terrain multiplier from movement_cost
+    let terrain_multiplier = terrain.base.movement_cost();
+
+    // Affinity modifier: 0.8 if tribute has affinity for this terrain, else 1.0
+    let affinity_modifier = if tribute.terrain_affinity.contains(&terrain.base) {
+        0.8
+    } else {
+        1.0
+    };
+
+    // Desperation multiplier: 1.0 + (0.5 × (1.0 - health%))
+    let health_percent = tribute.attributes.health as f32 / 100.0;
+    let desperation_multiplier = 1.0 + (0.5 * (1.0 - health_percent));
+
+    // Calculate final cost with all multipliers
+    let final_cost = base_cost * terrain_multiplier * affinity_modifier * desperation_multiplier;
+
+    // Round to nearest integer
+    final_cost.round() as u32
+}
